refactor: review-comment: calculate the colors on the frontend
As we otherwise would be mixing frontend and backend development

So that the most critical files appear a the top
