+++ src/Frontend/test-helpers/general-test-helpers.ts
@@ -15,15 +15,15 @@ import {
 } from '../../shared/shared-types';
 import {
   EMPTY_FREQUENT_LICENSES,
-  EMPTY_PROJECT_CONFIG,
   EMPTY_PROJECT_METADATA,
+  EMPTY_RAW_PROJECT_CONFIG,
 } from '../shared-constants';
 import { canResourceHaveChildren } from '../util/can-resource-have-children';
 
 const EMPTY_PARSED_FILE_CONTENT: ParsedFileContent = {
   metadata: EMPTY_PROJECT_METADATA,
   resources: {},
-  config: EMPTY_PROJECT_CONFIG,
+  config: EMPTY_RAW_PROJECT_CONFIG,
   manualAttributions: {
     attributions: {},
     resourcesToAttributions: {},
